rag/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4
ollama==0.1.6
pydantic==2.5.0
numpy==1.24.3